Clean up PlaceSelect a tiny bit
